requirements.txt
matplotlib==3.5.1
numpy==1.21.5
pandas==1.3.5
torch==1.10.1
cookiecutter==1.7.3
python_dotenv==0.19.2
seaborn==0.11.2